requirements.txt
cryptography==46.0.1
fastapi==0.117.1
google_api_python_client==2.167.0
httpx==0.28.1
langchain==0.3.27
langchain_core==0.3.76
langchain_openai==0.3.33
langgraph==0.6.7
langsmith==0.4.31
nest_asyncio==1.6.0
numpy==2.3.3
pandas==2.3.2
protobuf==6.32.1
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
Requests==2.32.5
scikit_learn==1.7.2
sentence_transformers==4.1.0
streamlit==1.45.0
torch==2.7.0
urllib3==2.5.0
